+++ CHANGELOG.md
@@ -6,6 +6,11 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-5452 Fixed not working cluster deployment GUI after SP3 changes
+* VFS-5425 Service Pack 3 for unified GUI
+* VFS-5411 Basic-auth sign-in changed to emergency passphrase
+* VFS-5377 Simple members view in emergency mode of Onepanel GUI
+* VFS-4640 Added auto-polling of space occupancy, disabled join space in emergency, changed session expired message
 * VFS-4596 Major changes to support unified and standalone Onepanel GUI
 * VFS-5242 Using new async backend to show progress of saving file popularity settings
 * VFS-5187 Fixed some graphical issues in supported spaces view; onepanel client library update